[FIX} mission 이름 오타 수정
