--- src/main/java/com/example/umc10th/domain/member/controller/MemberController.java
@@ -1,4 +1,0 @@
-package com.example.umc10th.domain.member.controller;
-
-public class MemberController {
-}
