--- src/main/java/com/example/umc10th/domain/member/converter/MissionConverter.java
@@ -1,4 +1,0 @@
-package com.example.umc10th.domain.member.converter;
-
-public class MissionConverter {
-}
