--- src/main/java/com/example/umc10th/domain/member/dto/MissionReqDTO.java
@@ -1,4 +1,0 @@
-package com.example.umc10th.domain.member.dto;
-
-public class MissionReqDTO {
-}
