--- src/main/java/com/example/umc10th/domain/member/entity/Store.java
@@ -1,4 +1,0 @@
-package com.example.umc10th.domain.member.entity;
-
-public class Store {
-}
