--- src/main/java/com/example/umc10th/domain/member/entity/mapping/MemberMission.java
@@ -1,4 +1,0 @@
-package com.example.umc10th.domain.member.entity.mapping;
-
-public class MemberMission {
-}
