--- src/main/java/com/example/umc10th/domain/member/enums/Address.java
@@ -1,4 +1,0 @@
-package com.example.umc10th.domain.member.enums;
-
-public enum Address {
-}
