--- src/main/java/com/example/umc10th/domain/member/exception/code/MissionErrorCode.java
@@ -1,4 +1,0 @@
-package com.example.umc10th.domain.member.exception.code;
-
-public enum MissionErrorCode {
-}
