--- src/main/java/com/example/umc10th/domain/member/exception/code/MissionSuccessCode.java
@@ -1,4 +1,0 @@
-package com.example.umc10th.domain.member.exception.code;
-
-public class MissionSuccessCode {
-}
