--- src/main/java/com/example/umc10th/domain/member/repository/MissionRepository.java
@@ -1,4 +1,0 @@
-package com.example.umc10th.domain.member.repository;
-
-public interface MissionRepository {
-}
