--- src/main/java/com/example/umc10th/domain/member/service/MissionService.java
@@ -1,4 +1,0 @@
-package com.example.umc10th.domain.member.service;
-
-public class MissionService {
-}
