+++ src/main/java/com/example/umc10th/domain/mission/controller/MissionrController.java
@@ -1,0 +1,4 @@
+package com.example.umc10th.domain.mission.controller;
+
+public class MissionrController {
+}
