+++ src/main/java/com/example/umc10th/domain/mission/converter/MissionConverter.java
@@ -1,0 +1,4 @@
+package com.example.umc10th.domain.mission.converter;
+
+public class MissionConverter {
+}
