+++ src/main/java/com/example/umc10th/domain/mission/dto/MissionResDTO.java
@@ -1,0 +1,4 @@
+package com.example.umc10th.domain.mission.dto;
+
+public class MissionResDTO {
+}
