+++ src/main/java/com/example/umc10th/domain/mission/entity/mapping/MemberMission.java
@@ -1,0 +1,4 @@
+package com.example.umc10th.domain.mission.entity.mapping;
+
+public class MemberMission {
+}
