+++ src/main/java/com/example/umc10th/domain/mission/enums/Address.java
@@ -1,0 +1,4 @@
+package com.example.umc10th.domain.mission.enums;
+
+public enum Address {
+}
