+++ src/main/java/com/example/umc10th/domain/mission/exception/MissionException.java
@@ -1,4 +1,4 @@
-package com.example.umc10th.domain.member.exception;
+package com.example.umc10th.domain.mission.exception;
 
 public class MissionException extends RuntimeException {
     public MissionException(String message) {
