+++ src/main/java/com/example/umc10th/domain/mission/exception/code/MissionErrorCode.java
@@ -1,0 +1,4 @@
+package com.example.umc10th.domain.mission.exception.code;
+
+public enum MissionErrorCode {
+}
