+++ src/main/java/com/example/umc10th/domain/mission/repository/MissionRepository.java
@@ -1,0 +1,4 @@
+package com.example.umc10th.domain.mission.repository;
+
+public interface MissionRepository {
+}
