+++ src/main/java/com/example/umc10th/domain/mission/service/MissionService.java
@@ -1,0 +1,4 @@
+package com.example.umc10th.domain.mission.service;
+
+public class MissionService {
+}
